refactor : 마커 애니메이션을 관리하는 컨트롤러 분리
- UIViewController에서 마커 애니메이션을 관리하는 컨트롤러를 분리하여 책임 완화
